requirements.txt
werkzeug
flask
flask_restful
openstacksdk
uwsgi; sys_platform != 'win32'
